move reflog commit loader into loaders package
